VideoDb: Prevent unwanted pagination
The last video includes a ref to force pagination when it comes
into view. This is achieved by watching the element (not the ref).

If the filter means that element comes into view, pagination will happen
automatically, unless that element is replaced by React.

Extend the key of the element to force re-render if it's no longer the
last element.

Resolves #76